pyproject.toml
[project]
name = "python-charuco"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.4",
    "opencv-contrib-python>=4.11.0.86",
    "pillow>=12.0.0",
    "reportlab>=4.4.4",
]
